pyproject.toml
[project]
name = "powerpointkaraoke"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv==1.0.0",
    "requests==2.31.0",
    "tqdm>=4.67.1",
]
